* device-src/activate-devpay.c: fix type warning
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@907 a8d146d6-cc15-0410-8900-af154a0219e0

+++ device-src/activate-devpay.c
@@ -147,7 +147,7 @@ static void parser_got_text(GMarkupParseContext * context,
             /* Do nothing; wait for the message. */
         }
     } else if (g_strrstr(current_tag, "Message")) {
-        g_set_error(error, G_MARKUP_ERROR, -1, "%.*s", text_len, text);
+        g_set_error(error, G_MARKUP_ERROR, -1, "%.*s", (int)text_len, text);
     }
 }
